# faster mulMonty