--- a/ocl/kernel2.cl
+++ b/ocl/kernel2.cl
@@ -38,7 +38,9 @@ INLINE uint _subMod(const uint lhs, const uint rhs, const uint p)
 // If lhs = x and rhs = y * 2^32 then r = x * y mod p.
 INLINE uint _mulMonty(const uint lhs, const uint rhs, const uint p, const uint q)
 {
-	const uint t_lo = lhs * rhs, t_hi = mul_hi(lhs, rhs);
+	const ulong prod = (ulong)lhs * rhs;
+	const uint t_lo = (uint)prod;
+	const uint t_hi = prod >> 32;
 	const uint mp = mul_hi(t_lo * q, p);
 	return _subMod(t_hi, mp, p);
 }
@@ -954,7 +956,7 @@ INLINE uint barrett(const ulong a, const uint b, const uint b_inv, const int b_s
 
 	const uint d = mul_hi((uint)(a >> b_s), b_inv), r = (uint)a - d * b;
 	const bool o = (r >= b);
-	*a_p = o ? d + 1 : d;
+	*a_p = d + (uint)o;
 	return o ? r - b : r;
 }