about 5% faster on Radeon 7900 XTX

before
```
400000000^{2^16} + 1: 00:00:49, 0.0262 ms/bit, data size: 2.25 MB.
250000000^{2^17} + 1: 00:02:14, 0.0369 ms/bit, data size: 4.5 MB.
40000000^{2^18} + 1: 00:06:07, 0.0555 ms/bit, data size: 9 MB.
9000000^{2^19} + 1: 00:21:46, 0.108 ms/bit, data size: 18 MB.
3500000^{2^20} + 1: 01:18:42, 0.207 ms/bit, data size: 36 MB.
1500000^{2^21} + 1: 03:17:50, 0.276 ms/bit, data size: 48 MB.
400000^{2^22} + 1: 11:52:54, 0.548 ms/bit, data size: 96 MB.
1200000^{2^22} + 1: 12:52:49, 0.547 ms/bit, data size: 96 MB.
500000^{2^23} + 1: 51:10:24, 1.16 ms/bit, data size: 192 MB.
```

after
```
400000000^{2^16} + 1: 00:00:47, 0.0254 ms/bit, data size: 2.25 MB.
250000000^{2^17} + 1: 00:02:11, 0.036 ms/bit, data size: 4.5 MB.
40000000^{2^18} + 1: 00:05:46, 0.0523 ms/bit, data size: 9 MB.
9000000^{2^19} + 1: 00:20:52, 0.103 ms/bit, data size: 18 MB.
3500000^{2^20} + 1: 01:15:19, 0.198 ms/bit, data size: 36 MB.
1500000^{2^21} + 1: 03:11:20, 0.267 ms/bit, data size: 48 MB.
400000^{2^22} + 1: 11:30:24, 0.531 ms/bit, data size: 96 MB.
1200000^{2^22} + 1: 12:28:20, 0.53 ms/bit, data size: 96 MB.
500000^{2^23} + 1: 49:16:19, 1.12 ms/bit, data size: 192 MB.
```