@galloty 2025-02-10
Your modification is the previous version of the code (23.07.0, which is the version in production on PrimeGrid):
```
inline uint REDC(const ulong t, const uint p, const uint q)
{
	const uint mp = mul_hi((uint)(t) * q, p), t_hi = (uint)(t >> 32), r = t_hi - mp;
	return (t_hi < mp) ? r + p : r;
}

inline uint _mulMonty(const uint lhs, const uint rhs, const uint p, const uint q)
{
	return REDC(lhs * (ulong)(rhs), p, q);
}
```

The new version
```
inline uint _mulMonty(const uint lhs, const uint rhs, const uint p, const uint q)
{
	const uint t_lo = lhs * rhs, t_hi = mul_hi(lhs, rhs);
	const uint mp = mul_hi(t_lo * q, p);
	return _subMod(t_hi, mp, p);
}
```
is a bit faster on some Nvidia GPU. The reason is that a radix-4 butterfly calls `_mulMonty(z1, w1, p, q)` and `_mulMonty(z2, w1, p, q)`, where `p` and `q` are constant. We have _t_lo = z1 * w1_ and _mp = mul_hi(t_lo * q, p)_. The compiler generates _t =  w1 * q_ once and _mp1 = mul_hi(z1 * t, p)_, _mp2 = mul_hi(z2 * t, p)_. If a "mul_wide" operation is used (`t = lhs * (ulong)(rhs)`), this optimization is not possible.

In practice, the throughput of mul and mul_hi is one cycle and of mul_wide is two cycles. Then both versions of the code are expected to be equivalent, except the optimization I stated above. Note that the latest versions of compilers may generate a mul_wide operation with the `t_lo = lhs * rhs, t_hi = mul_hi(lhs, rhs);` syntax and vice versa. It is quite surprising if AMD OpenCL compiler of RX 7000 Series makes a difference between `t_lo = lhs * rhs, t_hi = mul_hi(lhs, rhs)` and `ulong t = lhs * (ulong)(rhs); uint t_lo = (uint)(t), t_hi = (uint)(t >> 32);` Often, the code is faster not because of the mul_wide / mul+mul_hi  instructions but because the variable z1, z2, z3 and z4 are accessed in a different order. It would be interesting to examine asm code by setting the _ocl_debug_ flag.

Today, every compiler will produce the same machine code for `b ? 1 : 0` and `(int)b` and the first version is clearer for a human reader.