@ahorek 2025-02-10
thanks for your input, here's the difference:

_/opt/rocm/bin/amdclang -x cl -target amdgcn-amd-amdhsa -mcpu=gfx1100 -O3 -S -emit-llvm kernel.cl -o kernel.ll_
![obrazek](https://github.com/user-attachments/assets/b8935c44-50f5-4c5a-9601-22fa37295728)

I haven't tested it on NVIDIA, but if there's no improvement, it likely isn't worth it. I attempted to rewrite it using shifts, but requiring 128-bit vectors would negate any performance gains.